refactor:#39 BaseUrl 적용
