+++ src/components/detail/Info.tsx
@@ -2,7 +2,8 @@ import { Movie } from '@/types/DetailMovie';
 import { formatOverview } from '@/utils/formatFunction';
 import Image from 'next/image';
 import WrapperBox from '@/components/detail/WrapperBox';
-import { TMDB_IMG_URL } from '@/constants/tmdbConstants';
+// import { TMDB_IMG_URL } from '@/constants/tmdbConstants';
+import { TMDB_IMG_URL } from '@/constants/tmdbBaseUrl';
 
 interface Props {
   movie: Movie;
@@ -51,7 +52,12 @@ const Info = ({ movie }: Props) => {
             return (
               <div key={company.id}>
                 {company.logo_path ? (
-                  <Image src={TMDB_IMG_URL + company.logo_path} width={200} height={100} alt={company.name} />
+                  <Image
+                    src={TMDB_IMG_URL + '/t/p/w300' + company.logo_path}
+                    width={200}
+                    height={100}
+                    alt={company.name}
+                  />
                 ) : (
                   <p>{company.name}</p>
                 )}
